Created utility.js file and imported it
